feat: 파일 관리 및 UI State 관리
객체 분리
- 객체들의 의존성 분리 및 패키지 별 관리

UI State 관리
- UI State 관리르 위해 ViewModel에 변경 관리로 개선
