+++ app/src/main/java/com/example/todo/airbnb/data/model/date/CalendarYear.kt
@@ -1,0 +1,3 @@
+package com.example.todo.airbnb.data.model.date
+
+typealias CalendarYear = List<CalendarMonth>
