+++ app/src/main/java/com/example/todo/airbnb/data/model/date/DatesSelected.kt
@@ -1,8 +1,9 @@
 package com.example.todo.airbnb.presentation.search.date.components
 
-import com.example.todo.airbnb.data.datasource.CalendarYear
-import com.example.todo.airbnb.data.datasource.MonthOfYear
-import java.util.*
+import com.example.todo.airbnb.data.model.date.CalendarMonth
+import com.example.todo.airbnb.data.model.date.CalendarYear
+import com.example.todo.airbnb.data.model.date.DayOfWeek
+import com.example.todo.airbnb.data.model.date.MonthOfYear
 
 data class DaySelected(
     val day: Int,
@@ -24,7 +25,9 @@ data class DaySelected(
             year.indexOf(other.month)
         )
     }
-}
 
-val DaySelectedEmpty =
-    DaySelected(0, CalendarMonth("", "", 0, 0, DayOfWeek.Sunday), emptyList())
+    companion object {
+        val DaySelectedEmpty =
+            DaySelected(0, CalendarMonth("", "", 0, 0, DayOfWeek.Sunday), emptyList())
+    }
+}
