+++ app/src/main/java/com/example/todo/airbnb/data/repository/DateeRepositoryImpl.kt
@@ -1,34 +1,25 @@
 package com.example.todo.airbnb.data.repository
 
-import com.example.todo.airbnb.data.datasource.CalendarYear
-import com.example.todo.airbnb.data.datasource.DatesLocalDataSource
+import com.example.todo.airbnb.data.model.date.CalendarMonth
+import com.example.todo.airbnb.data.model.date.CalendarYear
+import com.example.todo.airbnb.data.model.date.DayOfWeek
+import com.example.todo.airbnb.data.model.date.MonthOfYear
 import com.example.todo.airbnb.domain.repository.DateRepository
-import com.example.todo.airbnb.presentation.search.date.components.DatesSelectedState
-import com.example.todo.airbnb.presentation.search.date.components.DaySelected
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
+import java.time.LocalDate
+import java.time.temporal.TemporalAdjusters
 
-class DateRepositoryImpl(
-    private val datesLocalDataSource: DatesLocalDataSource,
-) : DateRepository {
+class DateRepositoryImpl() : DateRepository {
 
     override fun getDate(): CalendarYear {
-        return datesLocalDataSource.calendarYear
-    }
-
-    override fun getDatesSelected(): DatesSelectedState {
-        return datesLocalDataSource.datesSelected
-    }
-
-    override suspend fun onDaySelected(daySelected: DaySelected) {
-        withContext(Dispatchers.IO) {
-            datesLocalDataSource.onDaySelected(daySelected)
-        }
-    }
-
-    override suspend fun onClear() {
-        withContext(Dispatchers.IO) {
-            datesLocalDataSource.onClear()
+        return MonthOfYear.values().map {
+            val date = LocalDate.of(2021, it.order, 1)
+            CalendarMonth(
+                name = it.name,
+                year = date.year.toString(),
+                numDays = date.with(TemporalAdjusters.lastDayOfMonth()).dayOfMonth,
+                monthNumber = it.order,
+                startDayOfWeek = DayOfWeek.values(date.dayOfWeek.toString()) ?: DayOfWeek.Sunday
+            )
         }
     }
 }
