+++ app/src/main/java/com/example/todo/airbnb/domain/repository/DateRepository.kt
@@ -1,15 +1,8 @@
 package com.example.todo.airbnb.domain.repository
 
-import com.example.todo.airbnb.data.datasource.CalendarYear
-import com.example.todo.airbnb.presentation.search.date.components.DatesSelectedState
-import com.example.todo.airbnb.presentation.search.date.components.DaySelected
+import com.example.todo.airbnb.data.model.date.CalendarMonth
 
 interface DateRepository {
-    fun getDate(): CalendarYear
 
-    fun getDatesSelected(): DatesSelectedState
-
-    suspend fun onDaySelected(daySelected: DaySelected)
-
-    suspend fun onClear()
+    fun getDate(): List<CalendarMonth>
 }
