+++ app/src/main/java/com/example/todo/airbnb/presentation/main/MainActivity.kt
@@ -1,11 +1,9 @@
 package com.example.todo.airbnb.presentation.main
 
-import android.os.Build
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
-import androidx.annotation.RequiresApi
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.material.MaterialTheme
