+++ app/src/main/java/com/example/todo/airbnb/presentation/search/date/components/DaySelectedStatus.kt
@@ -1,0 +1,5 @@
+package com.example.todo.airbnb.presentation.search.date.components
+
+enum class DaySelectedStatus {
+    NoSelected, Selected, NonClickable, FirstDay, LastDay, FirstLastDay
+}
